rustc_metadata: Merge `cstore.rs` into `creader.rs`
